move old main.rs to legacy_main.rs and keep new main.rs small
